+++ tests/io/test_postgres.py
@@ -2,6 +2,7 @@ from typing import Tuple
 from unittest import mock
 
 import apache_beam as beam
+import apache_beam.transforms.window as window
 from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.testing.util import assert_that, equal_to
 from psycopg import Error, IntegrityError
@@ -32,9 +33,6 @@ class TestWriteToPostgres:
 
     @mock.patch("psycopg.connect")
     def test_if_write_errors_are_returned(self, mock_connect: mock.Mock):
-        def equal_fn(expected: Tuple[Tuple[str], Error], actual: Tuple[Tuple[str], Error]):
-            return expected[0] == actual[0] and isinstance(actual[1], type(expected[1]))
-
         mock_connect.return_value.cursor.return_value.__enter__.return_value.execute.side_effect = IntegrityError()
 
         with TestPipeline() as p:
@@ -42,4 +40,22 @@ class TestWriteToPostgres:
             errors = data | "Writing example records to database" >> WriteToPostgres(
                 "host=localhost dbname=examples user=postgres password=postgres", "dummy", 2
             )
-            assert_that(errors, equal_to([(("test_1",), IntegrityError())], equal_fn))
+            assert_that(errors, equal_to([(("test_1",), IntegrityError())], failed_rows_equal_fn))
+
+    @mock.patch("psycopg.connect")
+    def test_if_works_with_non_global_windows(self, mock_connect: mock.Mock):
+        mock_connect.return_value.cursor.return_value.__enter__.return_value.execute.side_effect = IntegrityError()
+
+        with TestPipeline() as p:
+            errors = (
+                p
+                | "Reading example records" >> beam.Create([("test_1",)])
+                | beam.WindowInto(window.FixedWindows(1))
+                | "Writing example records to database"
+                >> WriteToPostgres("host=localhost dbname=examples user=postgres password=postgres", "dummy")
+            )
+            assert_that(errors, equal_to([(("test_1",), IntegrityError())], failed_rows_equal_fn))
+
+
+def failed_rows_equal_fn(expected: Tuple[Tuple[str], Error], actual: Tuple[Tuple[str], Error]):
+    return expected[0] == actual[0] and isinstance(actual[1], type(expected[1]))